requirements.txt
asgiref==3.2.10
Django==3.1
djangorestframework==3.11.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-ldap @ file:///C:/pythonArena/offlinePaackages/pyton-ldap/v3.2.0/python_ldap-3.2.0-cp38-cp38-win32.whl
pytz==2020.1
sqlparse==0.3.1
